+++ contracts/hackatom/tests/integration.rs
@@ -171,14 +171,7 @@ fn sudo_can_steal_tokens() {
     let res: Response = sudo(&mut deps, mock_env(), sys_msg).unwrap();
     assert_eq!(1, res.messages.len());
     let msg = res.messages.get(0).expect("no message");
-    assert_eq!(
-        msg,
-        &BankMsg::Send {
-            to_address: to_address.into(),
-            amount
-        }
-        .into(),
-    );
+    assert_eq!(msg, &BankMsg::Send { to_address, amount }.into(),);
 }
 
 #[test]
@@ -248,7 +241,7 @@ fn execute_release_works() {
     assert_eq!(
         msg,
         &BankMsg::Send {
-            to_address: beneficiary.into(),
+            to_address: beneficiary,
             amount: coins(1000, "earth"),
         }
         .into(),