+++ packages/std/src/ibc.rs
@@ -7,7 +7,6 @@ use serde::{Deserialize, Serialize};
 use std::cmp::{Ord, Ordering, PartialOrd};
 use std::fmt;
 
-use crate::addresses::HumanAddr;
 use crate::binary::Binary;
 use crate::coins::Coin;
 use crate::results::{Attribute, CosmosMsg, Empty, SubMsg};
@@ -27,7 +26,7 @@ pub enum IbcMsg {
         /// exisiting channel to send the tokens over
         channel_id: String,
         /// address on the remote chain to receive these tokens
-        to_address: HumanAddr,
+        to_address: String,
         /// packet data only supports one coin
         /// https://github.com/cosmos/cosmos-sdk/blob/v0.40.0/proto/ibc/applications/transfer/v1/transfer.proto#L11-L20
         amount: Coin,